# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: API Aluno
  description: Exemplo de gerenciamento de alunos
  contact:
    name: Caio Vinicius
    url: https://con6cao.github.io/portifolium/
    email: caio.dev.conceicao@gmail.com
  x-logo:
    url: http://localhost:8080/logo.png
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /aluno/{id}:
    get:
      tags:
      - Cadastro para Aluno
      summary: Buscar o Aluno
      description: Todos os cadastro de alunos serão feitos por essa rota
      operationId: buscarById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UsuarioDTO"
    put:
      tags:
      - Cadastro para Aluno
      summary: Atualizar Cadastro do Alunos
      operationId: atualizar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UsuarioDTO"
    post:
      tags:
      - Cadastro para Aluno
      summary: Inserir  Aluno
      description: Todos os cadastro de alunos serão feitos por essa rota
      operationId: inserir
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    delete:
      tags:
      - Cadastro para Aluno
      summary: Deletar o Aluno
      operationId: deletar
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usuario"
  /acesso:
    get:
      tags:
      - view-controler
      operationId: viewMetods
      parameters:
      - name: msg
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /:
    get:
      tags:
      - Cadastro para Aluno
      summary: Buscar varios Alunos
      description: Todos os cadastro de alunos serão feitos por essa rota
      operationId: buscar_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UsuarioDTO"
components:
  schemas:
    UsuarioDTO:
      required:
      - idade
      - nome
      - nome_Professor
      - nt_Psem
      - nt_Ssem
      - num_Sala
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        idade:
          type: string
        nt_Psem:
          type: string
        nt_Ssem:
          type: string
        num_Sala:
          type: string
        nome_Professor:
          type: string
    Usuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        idade:
          type: string
        nt_Psem:
          type: string
        nt_Ssem:
          type: string
        num_Sala:
          type: string
        nome_Professor:
          type: string